modified project structure for serverless deployment
