--- apps/server/src/lambda.ts
@@ -1,4 +1,0 @@
-import serverlessExpress from '@codegenie/serverless-express';
-import app from './app';
-
-export const handler = serverlessExpress({ app });
